Move promise logic to route; simplify main-map
Main-map had a single hook that was used to load all the map data, including fetching it. Right now that is mostly the case, but the fetch is now happening in the application model where promise-resolution is something we get for free.

One drawback is that if any of the promises fail, the entire route transition will go to an error state. This would have been true in the previous API, but at least now there’s a way to give feedback to the user on error.
