+++ app/components/main-map.js
@@ -1,14 +1,12 @@
 import Ember from 'ember';
 import mapboxgl from 'mapbox-gl';
 import { computed } from 'ember-decorators/object'; // eslint-disable-line
-import { task } from 'ember-concurrency';
+import sources from '../sources';
 
-import carto from '../utils/carto2';
 import layerGroups from '../layer-groups';
 
 import highlightedLotLayer from '../layers/highlighted-lot';
 import selectedLayers from '../layers/selected-lot';
-import sources from '../sources';
 
 const selectedFillLayer = selectedLayers.fill;
 const selectedLineLayer = selectedLayers.line;
@@ -32,7 +30,8 @@ export default Ember.Component.extend({
   mapConfig: Object.keys(layerGroups).map(key => layerGroups[key]),
 
   loading: true,
-  sourcesLoaded: false,
+  sourcesLoaded: true,
+  cartoSources: [],
 
   @computed('mainMap.selected')
   isSelectedBoundsOptions(selected) {
@@ -83,27 +82,26 @@ export default Ember.Component.extend({
   selectedFillLayer,
   selectedLineLayer,
 
-  configWithTemplate(config) {
-    return this.get('templateTask').perform(config);
-  },
-
-  templateTask: task(function* (config) {
-    const { minzoom = 0 } = config;
-    return yield carto.getVectorTileTemplate(config['source-layers'])
-      .then(template => ({
-        id: config.id,
-        type: 'vector',
-        tiles: [template],
-        minzoom,
-      }));
-  }),
-
   actions: {
     handleMapLoad(map) {
       window.map = map;
       const mainMap = this.get('mainMap');
       mainMap.set('mapInstance', map);
 
+      // add carto sources
+      this.get('cartoSources').forEach((sourceConfig) => {
+        map.addSource(sourceConfig.id, sourceConfig);
+      });
+
+      // add raster sources
+      Object.keys(sources)
+        .filter(key => sources[key].type === 'raster')
+        .forEach((key) => {
+          const source = sources[key];
+          map.addSource(source.id, source);
+        });
+
+      // setup controls
       const geoLocateControl = new mapboxgl.GeolocateControl({
         positionOptions: {
           enableHighAccuracy: true,
@@ -118,31 +116,6 @@ export default Ember.Component.extend({
       map.addControl(new mapboxgl.NavigationControl(), 'top-left');
       map.addControl(new mapboxgl.ScaleControl({ unit: 'imperial' }), 'bottom-left');
       map.addControl(geoLocateControl, 'top-left');
-
-      // add carto sources
-      const cartoSourcePromises = Object.keys(sources)
-        .filter(key => sources[key].type === 'cartovector')
-        .map((key) => {
-          const source = sources[key];
-          return this.configWithTemplate(source)
-            .then((sourceConfig) => {
-              map.addSource(sourceConfig.id, sourceConfig);
-            });
-        });
-
-      Promise.all(cartoSourcePromises)
-        .then(() => {
-          this.set('sourcesLoaded', true);
-        });
-
-      // add raster sources
-      Object.keys(sources)
-        .filter(key => sources[key].type === 'raster')
-        .forEach((key) => {
-          const source = sources[key];
-          map.addSource(source.id, source);
-        });
-
 
       // get rid of default building layer
       map.removeLayer('building');
