+++ app/templates/components/main-map.hbs
@@ -14,102 +14,96 @@
   mapLoaded=(action 'handleMapLoad')
   as |map|}}
 
-  {{#if sourcesLoaded}}
+  {{map.layer-group for='threed-buildings' visible=qps.threed-buildings}}
 
-    {{!-- {{map.call 'addControl' controlOptions 'top-left'}} --}}
+  {{map.layer-group for='neighborhood-tabulation-areas' visible=qps.neighborhood-tabulation-areas}}
 
-    {{map.layer-group for='threed-buildings' visible=qps.threed-buildings}}
+  {{map.layer-group for='assemblydistricts' visible=qps.assemblydistricts}}
 
-    {{map.layer-group for='neighborhood-tabulation-areas' visible=qps.neighborhood-tabulation-areas}}
+  {{map.layer-group for='nysenatedistricts' visible=qps.nysenatedistricts}}
 
-    {{map.layer-group for='assemblydistricts' visible=qps.assemblydistricts}}
+  {{map.layer-group for='nyccouncildistricts' visible=qps.nyccouncildistricts}}
 
-    {{map.layer-group for='nysenatedistricts' visible=qps.nysenatedistricts}}
+  {{map.layer-group for='community-districts' visible=qps.community-districts}}
 
-    {{map.layer-group for='nyccouncildistricts' visible=qps.nyccouncildistricts}}
+  {{map.layer-group for='boroughs' visible=qps.boroughs}}
 
-    {{map.layer-group for='community-districts' visible=qps.community-districts}}
+  {{map.layer-group for='subway' visible=qps.subway}}
 
-    {{map.layer-group for='boroughs' visible=qps.boroughs}}
+  {{map.layer-group for='building-footprints' visible=qps.building-footprints}}
 
-    {{map.layer-group for='subway' visible=qps.subway}}
+  {{map.layer-group for='pluto' visible=qps.pluto}}
 
-    {{map.layer-group for='building-footprints' visible=qps.building-footprints}}
+  {{map.layer-group for='commercial-overlays' visible=qps.commercial-overlays}}
 
-    {{map.layer-group for='pluto' visible=qps.pluto}}
+  {{map.layer-group for='landmarks' visible=qps.landmarks}}
 
-    {{map.layer-group for='commercial-overlays' visible=qps.commercial-overlays}}
+  {{map.layer-group for='historic-districts' visible=qps.historic-districts}}
 
-    {{map.layer-group for='landmarks' visible=qps.landmarks}}
+  {{map.layer-group for='scenic-landmarks' visible=qps.scenic-landmarks}}
 
-    {{map.layer-group for='historic-districts' visible=qps.historic-districts}}
+  {{map.layer-group for='waterfront-access-plan' visible=qps.waterfront-access-plan}}
 
-    {{map.layer-group for='scenic-landmarks' visible=qps.scenic-landmarks}}
+  {{map.layer-group for='coastal-zone-boundary' visible=qps.coastal-zone-boundary}}
 
-    {{map.layer-group for='waterfront-access-plan' visible=qps.waterfront-access-plan}}
+  {{map.layer-group for='low-density-growth-mgmt-areas' visible=qps.low-density-growth-mgmt-areas}}
 
-    {{map.layer-group for='coastal-zone-boundary' visible=qps.coastal-zone-boundary}}
+  {{map.layer-group for='sidewalkcafes' visible=qps.sidewalkcafes}}
 
-    {{map.layer-group for='low-density-growth-mgmt-areas' visible=qps.low-density-growth-mgmt-areas}}
+  {{map.layer-group for='fresh' visible=qps.fresh}}
 
-    {{map.layer-group for='sidewalkcafes' visible=qps.sidewalkcafes}}
+  {{map.layer-group for='transit-zones' visible=qps.transit-zones}}
 
-    {{map.layer-group for='fresh' visible=qps.fresh}}
+  {{map.layer-group for='inclusionary-housing' visible=qps.inclusionary-housing}}
 
-    {{map.layer-group for='transit-zones' visible=qps.transit-zones}}
+  {{map.layer-group for='mandatory-inclusionary-housing' visible=qps.mandatory-inclusionary-housing}}
 
-    {{map.layer-group for='inclusionary-housing' visible=qps.inclusionary-housing}}
+  {{map.layer-group for='special-purpose-subdistricts' visible=qps.special-purpose-subdistricts}}
 
-    {{map.layer-group for='mandatory-inclusionary-housing' visible=qps.mandatory-inclusionary-housing}}
+  {{map.layer-group for='special-purpose-districts' visible=qps.special-purpose-districts}}
 
-    {{map.layer-group for='special-purpose-subdistricts' visible=qps.special-purpose-subdistricts}}
+  {{map.layer-group for='zoning-map-amendments-pending' visible=qps.zoning-map-amendments-pending}}
 
-    {{map.layer-group for='special-purpose-districts' visible=qps.special-purpose-districts}}
+  {{map.layer-group for='zoning-map-amendments' visible=qps.zoning-map-amendments}}
 
-    {{map.layer-group for='zoning-map-amendments-pending' visible=qps.zoning-map-amendments-pending}}
+  {{map.layer-group for='zoning-districts' visible=qps.zoning-districts}}
 
-    {{map.layer-group for='zoning-map-amendments' visible=qps.zoning-map-amendments}}
+  {{!-- BASEMAPS --}}
 
-    {{map.layer-group for='zoning-districts' visible=qps.zoning-districts}}
+  {{map.layer-group for='digital-tax-map' visible=qps.digital-tax-map}}
 
-    {{!-- BASEMAPS --}}
+  {{map.layer-group for='aerials-16' visible=qps.aerials-16}}
 
-    {{map.layer-group for='digital-tax-map' visible=qps.digital-tax-map}}
-
-    {{map.layer-group for='aerials-16' visible=qps.aerials-16}}
-
-    {{#if mapMouseover.highlightedLotFeatures}}
-      {{#map.source
-        sourceId='highlighted-lot'
-        options=mapMouseover.highlightedLotSource as |source|}}
-        {{source.layer
-          layer=highlightedLotLayer
-          before='waterway-label'}}
-      {{/map.source}}
-    {{/if}}
-
-    {{hover-tooltip}}
-
-    {{#if mainMap.selected}}
-      {{#map.source
-        sourceId='selected-lot'
-        options=selectedLotSource as |source|}}
-        {{source.layer
-          layer=selectedFillLayer
-          before='waterway-label'}}
-        {{source.layer
-          layer=selectedLineLayer
-          before='waterway-label'}}
-      {{/map.source}}
-    {{/if}}
-
-    {{#if mainMap.currentAddress}}
-      {{#map.source sourceId='currentAddress' options=mainMap.addressSource as |source|}}
-        {{source.layer layer=mainMap.pointLayer}}
-      {{/map.source}}
-    {{/if}}
+  {{#if mapMouseover.highlightedLotFeatures}}
+    {{#map.source
+      sourceId='highlighted-lot'
+      options=mapMouseover.highlightedLotSource as |source|}}
+      {{source.layer
+        layer=highlightedLotLayer
+        before='waterway-label'}}
+    {{/map.source}}
   {{/if}}
 
+  {{hover-tooltip}}
+
+  {{#if mainMap.selected}}
+    {{#map.source
+      sourceId='selected-lot'
+      options=selectedLotSource as |source|}}
+      {{source.layer
+        layer=selectedFillLayer
+        before='waterway-label'}}
+      {{source.layer
+        layer=selectedLineLayer
+        before='waterway-label'}}
+    {{/map.source}}
+  {{/if}}
+
+  {{#if mainMap.currentAddress}}
+    {{#map.source sourceId='currentAddress' options=mainMap.addressSource as |source|}}
+      {{source.layer layer=mainMap.pointLayer}}
+    {{/map.source}}
+  {{/if}}
 
   {{#if shouldFitBounds}}
     {{map.call 'fitBounds' mainMap.bounds isSelectedBoundsOptions}}
